review fixes: an unbounded reference, a July-era default, and a progress line that lied
Three issues from reviewing #90.

**The reference-length bound was fixed in one test and not the other.** `gen_rare_test`
learned that "longest dev utterance" overshoots — it produced 29.6 s clips and the model's
own ">20s degrades voice cloning" warning — but `gen_en_gb_offglide_test` kept the same
`max(len(text))` heuristic. It happens to be safe for en_gb (dev tops out at 7.3 s), but the
script takes `--ref_lang`, and zu_za would have selected a 23.2 s reference with en_us at
18.2 s. That penalty lands on EVERY checkpoint equally, so it flattens the very comparison
the reference exists to support.

Both now use one bounded picker in `gen_accept_test` (the module they already share) rather
than two copies of the rule: longest WITHIN 4-15 s, nearest to the band if nothing is inside.
Verified every reference across both tests is now in band, zu_za included.

**`gen_accept_test --adapter` defaulted to `checkpoints/checkpoint-3000`** — the JULY run,
older than v5, v6 and v7, and it still exists on disk, so an acceptance test invoked without
`--adapter` silently judged a model nobody was shipping. This is the same stale-pin bug
already fixed in extract_diff.py, apply_diff.py and publish_hf.py, and a default that is
never bumped is exactly how it comes back. Now tracks the current fine-tune.

**`preview-all.mjs` misreported its own progress.** `pairs.length - todo.length` counts the
languages `--only` filtered out, so a 3-pair run announced "609 already present" when those
609 had never been rendered. A resumable sweep whose progress line lies about what it holds
is worse than one with no progress line; the two counts are now reported separately.

Docstring examples updated from checkpoint-4000 to the checkpoint-6000 that produced the
shipped ipa_diff_v7.
